backend/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
Flask-Migrate>=4.0
Flask-Cors>=4.0
PyJWT>=2.8
PyMySQL>=1.1